pyproject.toml
[tool.poetry]
name = "sport-school"
version = "0.1.0"
description = ""
authors = ["kirill_barkhatov <k.s.barkhatov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
python-telegram-bot = "^21.9"
djangorestframework = "^3.15.2"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
ipython = "^8.29.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 119
